deleted old table
